% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverWK.R
\name{canopyCover}
\alias{canopyCover}
\title{Oblicz zwarcie poziome koron drzew}
\usage{
canopyCover(las_path, Z, RN, hgtbreak)
}
\arguments{
\item{las_path}{ścieżka do znormalizowanej chmury punktów}

\item{Z}{wektor zawierający wysokości znormalizowanej chmury punktów}

\item{RN}{wektor zawierający numer odbicia (echa) punktu}

\item{hgtbreak}{próg wysokości powyżej którego ma być liczone zwarcie}
}
\value{
wartość zwarcia poziomego koron drzew
}
\description{
Oblicz zwarcie poziome koron drzew
}
\examples{
laspath = "D:/Wojtek/Geomatyka/dane_mce/dzialka_lidar/2023.laz"
Z = as.name("Z")
RN = as.name("ReturnNumber")
canopyCover(laspath, Z, RN, 2)
}
