% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_dokumentacja.R
\docType{data}
\name{df}
\alias{df}
\title{Example data}
\format{
\subsection{\code{moj_df}}{

A data frame with 25 observations on 2 variables.
\describe{
\item{liczby}{numeric vector}
\item{litery}{charcter vector}
}
}
}
\source{
Manually created.
}
\usage{
df
}
\description{
Example data created manually for the purpose of this package.
}
\keyword{datasets}
