% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMaxZ_funkcja.R
\name{findMaxZ}
\alias{findMaxZ}
\title{Znajdź najwyższe drzewo na podstawie znormalizowanej chmury punktów}
\usage{
findMaxZ(dirPath, hgtTresh)
}
\arguments{
\item{dirPath}{ścieżka do katalogu z plikiem/plikami .las}

\item{hgtTresh}{próg wysokości, powyżej którego zapisywane mają być najwyższe drzewa}
}
\value{
plik ESRI Shapefile o punktowej geometrii, lokalizujący najwyższe drzewo
}
\description{
Znajdź najwyższe drzewo na podstawie znormalizowanej chmury punktów
}
