% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_temp.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatury}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{temperatura_f}{wektor zawierajacy wartosci temperatury w stopniach Fahrenheita}
}
\value{
wektor o takiej samej długości jak @param
}
\description{
Ta funkcja przelicza temperaturę wyrażoną w stopniach Fahrenheita na stopnie Celsjusza
}
\examples{
konwersja_temp(75)
konwersja_temp(110)
konwersja_temp(0)
konwersja_temp(c(0, 75, 110))
}
